Split widget: sort header cell
